requirements.txt
# requirements.txt
# ── Web Framework ──────────────────────────────
Flask==2.3.3
Werkzeug==3.0.3
gunicorn==21.2.0              # Production WSGI server

# ── Database ───────────────────────────────────
Flask-PyMongo==2.3.0
pymongo==4.6.1

# ── Authentication ──────────────────────────────
Flask-Dance[google]==7.0.1

# ── AI / ML ─────────────────────────────────────
google-generativeai
huggingface-hub==0.23.2
transformers>=4.41.0
accelerate>=0.30.0
torch>=2.2.0
torchvision
torchaudio
bitsandbytes

# ── NLP / Sentiment ──────────────────────────────
nltk

# ── Utilities ────────────────────────────────────
python-dotenv==1.0.1
requests
numpy<2

# ── Dev / Testing (install separately in CI) ──────
# pytest
# pytest-cov
# flake8